Fixed a small error w/ scenario one
I commented out the "if response == true" part in scenario 1 in scenario controller, because it was causing the actors to not respond to me. I think the "response" variable doesn't do what we think it does, and I don't think it ever becomes true. The scenario interaction is all working properly now.